+++ dd-java-agent/instrumentation/trace-annotation/trace-annotation.gradle
@@ -2,7 +2,7 @@ apply from: "${rootDir}/gradle/java.gradle"
 
 dependencies {
   compile project(':dd-trace-ot')
-  compile project(':dd-trace-annotations')
+  compile project(':dd-trace-api')
   compile project(':dd-java-agent:tooling')
 
   compile deps.bytebuddy
