+++ dd-trace-api/dd-trace-api.gradle
@@ -2,7 +2,7 @@ apply from: "${rootDir}/gradle/java.gradle"
 apply from: "${rootDir}/gradle/publish.gradle"
 apply from: "${rootDir}/gradle/jacoco.gradle"
 
-description = 'dd-trace-annotations'
+description = 'dd-trace-api'
 dependencies {
   compile deps.slf4j
   testCompile deps.junit
