+++ dd-trace-api/src/main/java/datadog/trace/api/Trace.java
