+++ dd-trace-ot/src/main/java/datadog/opentracing/DDTracer.java
@@ -6,11 +6,11 @@ import datadog.opentracing.decorators.DDDecoratorsFactory;
 import datadog.opentracing.propagation.Codec;
 import datadog.opentracing.propagation.HTTPCodec;
 import datadog.trace.api.DDTags;
-import datadog.trace.api.DDTraceConfig;
-import datadog.trace.api.Service;
-import datadog.trace.api.sampling.AllSampler;
-import datadog.trace.api.sampling.Sampler;
-import datadog.trace.api.writer.Writer;
+import datadog.trace.common.DDTraceConfig;
+import datadog.trace.common.Service;
+import datadog.trace.common.sampling.AllSampler;
+import datadog.trace.common.sampling.Sampler;
+import datadog.trace.common.writer.Writer;
 import io.opentracing.ActiveSpan;
 import io.opentracing.ActiveSpanSource;
 import io.opentracing.BaseSpan;
