+++ dd-trace-ot/src/main/java/datadog/trace/common/Service.java
@@ -1,4 +1,4 @@
-package datadog.trace.api;
+package datadog.trace.common;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
