+++ dd-trace-ot/src/main/java/datadog/trace/common/sampling/AllSampler.java
@@ -1,4 +1,4 @@
-package datadog.trace.api.sampling;
+package datadog.trace.common.sampling;
 
 import datadog.opentracing.DDBaseSpan;
