+++ dd-trace-ot/src/main/java/datadog/trace/common/writer/Writer.java
@@ -1,8 +1,8 @@
-package datadog.trace.api.writer;
+package datadog.trace.common.writer;
 
 import datadog.opentracing.DDBaseSpan;
-import datadog.trace.api.DDTraceConfig;
-import datadog.trace.api.Service;
+import datadog.trace.common.DDTraceConfig;
+import datadog.trace.common.Service;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
